% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{sampledata}
\alias{sampledata}
\title{Sample Dataset from a Typical IAT}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 220 rows and 6 columns.}
\usage{
data(sampledata)
}
\description{
This dataset contains sample data from a typically-structured IAT.
This IAT has 220 trials, and this particular dataset does not meet criteria for exclusion.
}
\examples{
data(sampledata)
}
\keyword{datasets}
